all *.c: hush compiler by passing unsigned for ‘%x’ and similar.
The printf(3) family expect an unsigned integer for ‘%x’ so ensure they
get it to stop the compiler warning that a signed is being passed.  Each
case was considered rather than just mechanically adding a cast to
‘unsigned’.

+++ sbr/pidstatus.c
@@ -59,7 +59,7 @@ pidstatus (int status, FILE *fp, char *cp)
     fputs(mesg, fp);
 
     if (num != -1) {
-        fprintf(fp, " %#x", num);
+        fprintf(fp, " %#x", (unsigned)num);
         if (lookup) {
             errno = 0;
             signame = strsignal(num);
